chore: rename ember-package entity to ember-app-package
- Renamed to align to dependnecy-graph APIs.
- Added TODO and escape hatch when discoverEmberServiceDepednencies
  tries to parse a typescript file.
  - When running this against a pilot repo it was failing.
- Fixed type in tests/migration-strategy
