+++ packages/migration-graph-ember/src/entities/ember-app-package.ts
